requirements.txt
flask
flask_cors
pyserial
